+++ src/components/TarSelect.vue
@@ -25,6 +25,9 @@ const classes = computed<string[]>(() => {
       classes.push("form-select-sm");
       break;
   }
+  if (props.status) {
+    classes.push(`is-${props.status}`);
+  }
   return classes;
 });