+++ src/components/TarTextarea.vue
@@ -27,6 +27,9 @@ const classes = computed<string[]>(() => {
       classes.push("form-control-sm");
       break;
   }
+  if (props.status) {
+    classes.push(`is-${props.status}`);
+  }
   return classes;
 });
 const height = computed<string | undefined>(() => {